// src/app/models/SearchParams.ts
export interface SearchParams{
    q: string;
    center: string;
    description: string;
    description_508: string;
    keywords: string;
    location: string;
    media_type: string;
    nasa_id: string;
    page: number;
    page_size: number;
    photographer: string;
    secondary_creator: string;
    title: string;
    year_start: string;
    year_end: string;
}